requirements.txt
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
alembic==1.14.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.5.2
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.13.4
bleach==6.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==45.0.4
cycler==0.12.1
Cython==3.1.2
dataset==1.6.2
datasets==2.21.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.14.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.6.1
gensim==4.3.3
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
h11==0.16.0
haversine==2.9.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
latex2mathml==3.77.0
lazy_loader==0.4
Levenshtein==0.25.1
lmdb==1.6.2
lxml==5.4.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdtex2html==1.3.1
mdurl==0.1.2
mistune==3.1.3
mmcv==2.2.0
mmengine==0.10.7
model-index==0.1.11
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
nltk==3.9.1
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
packaging==24.2
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.5.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==27.0.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.11.6
requests==2.28.2
rich==13.4.2
rouge-chinese==1.0.3
rpds-py==0.20.1
ruff==0.11.13
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.54
stack-data==0.6.3
starlette==0.44.0
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.2.1
tomlkit==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.28.0
typer==0.16.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.55
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.33.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarg==0.1.9
yarl==1.15.2
zhconv==1.4.3
zipp==3.20.2
